pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "svx"
version = "0.2.0"
description = "Simulate, Verify, Execute — a safety layer for coding agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "mcp>=1.0.0",
]

[project.scripts]
svx = "svx.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
